external/pflash: Add quiet rules and fix race
Adding quiet rules to make our output a bit cleaner. Building now looks
like this:

$ make
    LN  libflash
    LN  common
    LN  ccan
    CC  pflash.o
    CC  version.o
    LD  common-arch_flash.o
    CC  pflash

You can see the full build ouput by doing a "make V=1".

By doing this, we build fractionally faster, exposing arace condition
between running the make_version.sh script and the link existing for it.
As we run it when creating the variable, there is no way to ensure it
exists first.

Solved this by not creating the symlink and simply running
make_version.sh from the root.

Signed-off-by: Joel Stanley <joel@jms.id.au>