external/pflash: Fix makefile dependencies
When building under buildroot, libflash was being built before the
links:

 make[2]: *** No rule to make target 'libflash/libflash.c', needed by
 'libflash-libflash.o'.  Stop.
 make[2]: *** Waiting for unfinished jobs....
     LN  ccan
     LN  common
     LN  libflash

To reproduce this outside of buildroot, set PFLASH_VERSION to anything.
This is another race that is only exposed in certain conditions. By
describing the dependencies on the source files the build works again.

I think it's almost time to stop using symlinks.

Signed-off-by: Joel Stanley <joel@jms.id.au>