Fix asm-offsets generation
There is a difference between the asm-offsets.s generated by clang and GCC:

--- asm/asm-offsets.s_clang
+++ asm/asm-offsets.s_gcc

-#define CPUTHREAD_PIR 8 /* offsetof(struct cpu_thread, pir) */
- # 0 "" 2
-       .loc 1 38 0
- # 38 "asm/asm-offsets.c" 1
+       #define CPUTHREAD_PIR 8 /* offsetof(struct cpu_thread, pir) */
+       #NO_APP
+       .loc    1 38 2                  # asm/asm-offsets.c:38:2
+       #APP

This breaks make_offests.sh, which looks for #define at the start
of the line. This is unnecessary, as the only #defines that appear in the
file are "good" ones we want. So drop the start of line requirement.

Signed-off-by: Joel Stanley <joel@jms.id.au>