Makefile: Disable warnings to make clang happy
Clang doesn't like some of the warnings we have, so silence the ones we
know about in order to enable the build to succeed. These should be
investigated and removed in once the code issues are resolved.

Signed-off-by: Joel Stanley <joel@jms.id.au>
---
v2: Make stack frame max size larger, as clang-7 is even worse than 6