test: Update qemu arguments to use bmc simulator
THe qemu skiboot platform as of 8340a9642bba ("plat/qemu: use the common
OpenPOWER routines to initialize") uses the common aspeed BMC setup
routines. This means a BT interface is always set up, and if the
corresponding Qemu model is not present the timeout is 30 seconds.

It looks like this every time an IPMI message is sent:

 BT: seq 0x9e netfn 0x06 cmd 0x31: Maximum queue length exceeded
 BT: seq 0x9d netfn 0x06 cmd 0x31: Removed from queue
 BT: seq 0x9f netfn 0x06 cmd 0x31: Maximum queue length exceeded
 BT: seq 0x9e netfn 0x06 cmd 0x31: Removed from queue
 BT: seq 0xa0 netfn 0x06 cmd 0x31: Maximum queue length exceeded
 BT: seq 0x9f netfn 0x06 cmd 0x31: Removed from queue

Avoid this by adding the bmc simulator model to the Qemu powernv
machine.

Signed-off-by: Joel Stanley <joel@jms.id.au>